Updated the Installer and some other files.

+++ WindowsInstaller/BetterDiscordWI/panels/Panel2.cs
@@ -13,7 +13,6 @@ namespace BetterDiscordWI.panels
 {
     public partial class Panel2 : UserControl, IPanel
     {
-
         private String _dataPath, _tempPath;
         private Utils _utils;
 
@@ -31,13 +30,32 @@ namespace BetterDiscordWI.panels
             GetParent().btnCancel.Enabled = false;
 
             _utils = new Utils();
-
-            AppendLog("Killing Discord");
-            foreach (var process in Process.GetProcessesByName("Discord"))
+            if (GetParent().DiscordPath.Contains("Discord\\"))
             {
-                process.Kill();
+                AppendLog("Killing Discord");
+                foreach (var process in Process.GetProcessesByName("Discord"))
+                {
+                    process.Kill();
+                }
             }
-            
+            if (GetParent().DiscordPath.Contains("DiscordCanary\\"))
+            {
+                AppendLog("Killing DiscordCanary");
+                foreach (var process in Process.GetProcessesByName("DiscordCanary"))
+                {
+                    process.Kill();
+                }
+            }
+            if (GetParent().DiscordPath.Contains("DiscordPTB\\"))
+            {
+                AppendLog("Killing DiscordPTB");
+                foreach (var process in Process.GetProcessesByName("DiscordPTB"))
+                {
+                    process.Kill();
+                    break;
+                }
+            }
+
             CreateDirectories();
         }
 
@@ -88,14 +106,24 @@ namespace BetterDiscordWI.panels
 
         private void DeleteDirs()
         {
+            int errors = 0;
             Thread t = new Thread(() =>
             {
                 String dir = GetParent().DiscordPath + "\\resources\\app";
 
                 if (Directory.Exists(dir))
                 {
-                    AppendLog("Deleting " + dir);
-                    Directory.Delete(dir, true);
+                    try
+                    {
+                        AppendLog("Deleting " + dir);
+                        Directory.Delete(dir, true);
+                    }
+                    catch
+                    {
+                        AppendLog("Error Failed to Delete the '" + dir + "\\resources\\app' Directory.");
+                        errors = 1;
+                        Finalize(errors);
+                    }
                 }
 
                 while (Directory.Exists(dir))
@@ -123,14 +151,22 @@ namespace BetterDiscordWI.panels
 
                 Directory.Move(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\BetterDiscord\\temp\\BetterDiscordApp-stable", GetParent().DiscordPath + "\\resources\\node_modules\\BetterDiscord");
 
-                AppendLog("Extracting app.asar");
+                try
+                {
+                    AppendLog("Extracting app.asar");
+                    AsarArchive archive = new AsarArchive(GetParent().DiscordPath + "\\resources\\app.asar");
 
-                AsarArchive archive = new AsarArchive(GetParent().DiscordPath + "\\resources\\app.asar");
+                    AsarExtractor extractor = new AsarExtractor();
+                    extractor.ExtractAll(archive, GetParent().DiscordPath + "\\resources\\app\\");
 
-                AsarExtractor extractor = new AsarExtractor();
-                extractor.ExtractAll(archive, GetParent().DiscordPath + "\\resources\\app\\");
-
-                Splice();
+                    Splice();
+                }
+                catch
+                {
+                    AppendLog("Error Extracting app.asar: Newtonsoft.Json.dll might not be present in the Installer Folder. Installation cannot Continue.");
+                    errors = 1;
+                    Finalize(errors);
+                }
             });
             
             
@@ -242,7 +278,18 @@ namespace BetterDiscordWI.panels
 
             if (GetParent().RestartDiscord)
             {
-                Process.Start(GetParent().DiscordPath + "\\Discord.exe");
+                if (GetParent().DiscordPath.Contains("\\Discord\\"))
+                {
+                    Process.Start(GetParent().DiscordPath + "\\Discord.exe");
+                }
+                if (GetParent().DiscordPath.Contains("\\DiscordCanary\\"))
+                {
+                    Process.Start(GetParent().DiscordPath + "\\DiscordCanary.exe");
+                }
+                if (GetParent().DiscordPath.Contains("\\DiscordPTB\\"))
+                {
+                    Process.Start(GetParent().DiscordPath + "\\DiscordPTB.exe");
+                }
             }
         }
